Preprovision PVCs with data sources and other fixes (#351)
- Add `database.persistence.preprovisionVolumes` value that allows PVCs
  for all non-hotcopy SMs to be preprovisioned. This allows the SM
  statefulset to be scaled in the absence of the volume snapshots used
  to restore the database, since they do not appear in the
  volumeClaimTemplates section when `preprovisionVolumes=true`.
- Fix bug in backup_hooks.py which was causing failures when a non-root
  user was configured in the pod security context. In that case, it is
  not necessary to change ownership and permissions on the backup.txt
  file, since the user is either the same as the nuodb user or root.
- Update nuosm to allow scaling of the SM statefulset in the absence of
  volume snapshots. Now SMs with non-0 ordinals can skip the existing
  backup ID check in the case where no archive snapshots are found.
- Fix a bug exposed by the relaxed backup ID check, which is that if an
  archive object and info.json file is created but the SM is not
  actually initialized (i.e. no state.dat file is created), then we were
  incorrectly checking the backup ID on the newly-created archive on
  restart and preventing the SM from starting.
- Add test coverage for snapshot restore.

+++ stable/database/files/backup_hooks.py
@@ -46,8 +46,9 @@ def write_file(path, content):
     # default. Making the file group-writable ensures that it is accessible to
     # the nuodb user in OpenShift deployments where an arbitrary uid is used
     # with gid 0.
-    os.chown(path, 1000, 0)
-    os.chmod(path, mode=0o660)
+    if os.getuid() == 0:
+        os.chown(path, 1000, 0)
+        os.chmod(path, mode=0o660)
 
 
 def get_nuodb_pids():
@@ -211,7 +212,7 @@ def post_backup(backup_id, query):
     # Delete backup ID files and payload file
     if os.path.exists(ARCHIVE_BACKUP_ID_FILE):
         os.remove(ARCHIVE_BACKUP_ID_FILE)
-    if os.path.exists(JOURNAL_BACKUP_ID_FILE):
+    if JOURNAL_BACKUP_ID_FILE is not None and os.path.exists(JOURNAL_BACKUP_ID_FILE):
         os.remove(JOURNAL_BACKUP_ID_FILE)
     if os.path.exists(BACKUP_PAYLOAD_FILE):
         os.remove(BACKUP_PAYLOAD_FILE)
